working on tuning pose
the robot currently loses its shit anytime we try to tell it to do anything.  if i tell it to go straight it just spazes out.  tried to do a test with the november branch code and polarity and all is right.  there is something else going on, unsure as of now