Yaw+Outboard posAssn, Rud. Teleop, Hanging
Added simplistic enumeration for yaw and lift allowing for a single method that carries through strings to tell the lift what position it is targeting instead of having multiple methods.  This will be critical in coming years when we need efficiency to be high.  Tele-op attempts were added, we will see how oskar actually enjoys it, we can always just change this too, but we will see.  Hanging motor and class were added, will have to be tuned, if it even will be on the robot.

WAYPOINT:  Positions and Fanciness for Tele-Op